requirements.txt
asgiref
Django
django-coreapi
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
asgiref
pytz
sqlparse
tzdata
Pillow
pandas
setuptools>=58.0.0